refactor: 내가 좋아요한 회고 조회 api 클래스 변경
