+++ src/main/java/com/example/floud/service/MemoirLikeService.java
@@ -1,6 +1,8 @@
 package com.example.floud.service;
 
+import com.example.floud.dto.request.like.LikeMemoirListRequestDto;
 import com.example.floud.dto.request.like.LikeSaveRequestDto;
+import com.example.floud.dto.response.like.LikeMemoirListResponseDto;
 import com.example.floud.dto.response.like.LikeSaveResponseDto;
 import com.example.floud.entity.Memoir;
 import com.example.floud.entity.MemoirLike;
@@ -12,6 +14,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
@@ -43,5 +48,31 @@ public class MemoirLikeService {
                 .orElseThrow(()-> new IllegalArgumentException("해당 좋아요 정보가 존재하지 않습니다. memoir_like_id = "+memoir_like_id));
         memoirLikeRepository.delete(memoirLike);
         return memoir_like_id;
+    }
+
+    @Transactional
+    public List<LikeMemoirListResponseDto> getMemoirLike(Long user_id, LikeMemoirListRequestDto requestDto){
+
+        User user = userRepository.findById(user_id)
+                .orElseThrow(()-> new IllegalArgumentException("해당 회원 정보가 존재하지 않습니다. user_id = "+ user_id));
+
+        LocalDate like_date =  requestDto.getLikeDate();
+        LocalDate startDate = like_date.withDayOfMonth(1);
+        LocalDate endDate = like_date.withDayOfMonth(like_date.lengthOfMonth());
+
+        List<MemoirLike> memoirLikes = memoirLikeRepository.findByUserIdAndLikeDateBetween(user_id, startDate, endDate);
+        List<LikeMemoirListResponseDto> responseDtos = memoirLikes.stream()
+                .map(memoirLike -> {
+                    Memoir memoir = memoirLike.getMemoir();
+                    return LikeMemoirListResponseDto.builder()
+                            .memoir_id(memoir.getId())
+                            .title(memoir.getTitle())
+                            .createdAt(memoir.getCreatedAt())
+                            .memoir_like_id(memoirLike.getMemoir_like_id())
+                            .build();
+                })
+                .collect(Collectors.toList());
+
+        return responseDtos;
     }
 }
