+++ src/main/java/com/example/floud/service/MemoirService.java
@@ -43,31 +43,7 @@ public class MemoirService {
         return responseDto;
     }
 
-    @Transactional
-    public List<LikeMemoirListResponseDto> getMemoirLike(Long user_id, LikeMemoirListRequestDto requestDto){
 
-        User user = userRepository.findById(user_id)
-                .orElseThrow(()-> new IllegalArgumentException("해당 회원 정보가 존재하지 않습니다. user_id = "+ user_id));
-
-        LocalDate like_date =  requestDto.getLikeDate();
-        LocalDate startDate = like_date.withDayOfMonth(1);
-        LocalDate endDate = like_date.withDayOfMonth(like_date.lengthOfMonth());
-
-        List<MemoirLike> memoirLikes = memoirLikeRepository.findByUserIdAndLikeDateBetween(user_id, startDate, endDate);
-        List<LikeMemoirListResponseDto> responseDtos = memoirLikes.stream()
-                .map(memoirLike -> {
-                    Memoir memoir = memoirLike.getMemoir();
-                    return LikeMemoirListResponseDto.builder()
-                            .memoir_id(memoir.getId())
-                            .title(memoir.getTitle())
-                            .createdAt(memoir.getCreatedAt())
-                            .memoir_like_id(memoirLike.getMemoir_like_id())
-                            .build();
-                })
-                .collect(Collectors.toList());
-
-        return responseDtos;
-    }
 
     @Transactional
     public MemoirUpdateResponseDto updateMemoir(Long memoirId, MemoirUpdateRequestDto requestDto) {
